Se paso login con google a LoginActivity
Se paso login con google a LoginActivity junto con el splash ademas de solucionar pequeños errores de este
