+++ app/src/main/AndroidManifest.xml
@@ -16,17 +16,10 @@
         android:theme="@style/Theme.RefactoringLife4"
         tools:targetApi="31">
         <activity
-            android:name=".LoginActivity"
-            android:exported="false">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
             android:name=".RegisterActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".LoginActivity"
             android:exported="true"
             android:theme="@style/SplashScreen">
             <intent-filter>
