--- app/src/main/java/com/example/refactoringlife4/MainActivity.kt
@@ -1,84 +1,0 @@
-package com.example.refactoringlife4
-
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.widget.Toast
-import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import com.example.refactoringlife4.databinding.ActivityMainBinding
-import com.google.android.gms.auth.api.signin.GoogleSignIn
-import com.google.android.gms.auth.api.signin.GoogleSignInOptions
-import com.google.android.gms.common.api.ApiException
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.GoogleAuthProvider
-
-class MainActivity : AppCompatActivity() {
-
-    private val GOOGLE_SING_IN = 100
-    val screenSplash = installSplashScreen()
-    private lateinit var binding: ActivityMainBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-        screenSplash.setKeepOnScreenCondition { true }
-        setup()
-    }
-
-    private fun setup() {
-
-        binding.root.setOnClickListener {
-            val googleConf = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-                .requestIdToken(getString(R.string.default_web_client_id))
-                .requestEmail()
-                .build()
-
-            val googleClient = GoogleSignIn.getClient(this, googleConf)
-            googleClient.signOut()
-            startActivityForResult(googleClient.signInIntent, GOOGLE_SING_IN)
-        }
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (requestCode == GOOGLE_SING_IN) {
-            val task = GoogleSignIn.getSignedInAccountFromIntent(data)
-
-            try {
-                val account = task.getResult(ApiException::class.java)
-                if (account != null) {
-                    account.getIdToken()?.let { idToken ->
-                        val credential = GoogleAuthProvider.getCredential(idToken, null)
-                        FirebaseAuth.getInstance().signInWithCredential(credential)
-                            .addOnCompleteListener {
-                                if (it.isSuccessful) {
-                                    Toast.makeText(
-                                        this,
-                                        "Logeado correctamente",
-                                        Toast.LENGTH_SHORT
-                                    ).show()
-                                } else {
-                                    Toast.makeText(
-                                        this,
-                                        "Error en la autenticación",
-                                        Toast.LENGTH_SHORT
-                                    ).show()
-                                }
-                            }
-                    } ?: run {
-                        Toast.makeText(
-                            this,
-                            "El token de autenticación es nulo",
-                            Toast.LENGTH_SHORT
-                        ).show()
-                    }
-                }
-            } catch (e: ApiException) {
-                Toast.makeText(this, "No se obtuvo un correo", Toast.LENGTH_SHORT).show()
-            }
-
-        }
-    }
-
-}
